requirements.txt
aiogram>=3.0
aiohttp
python-dotenv
apscheduler
pandas
openpyxl
aiosqlite
matplotlib
seaborn